feat: show assignment policy name in auto-assignment activity messages (#13598)
